Update analysis dependencies (#1174)
* deps: update analysis dependencies

This updates to Kotlin 1.4.0 and also updates a couple plugins.
Everything is fixed in a separate commit.

All versions are from now on in the `gradle.properties` file instead
of being distributed between the different gradle files.

* refactor: remove unused code and fix types

Some code paths could never be reached due to always evaluating to
e.g., a falsy value. Other code never used some of the arguments
passed through. This is fixed in this commit as well as formatting
the files along the new ktlint rules.

* refactor: apply ktlint auto formatting

This formats our analysis files according to the current ktlint rules.