tech: stop cloning objects in the reducer
The reducers should not require cloning of the objects as long as
the implementation works as expected. There's only one reducer that
still requires the cloning due to directly manipulating a state
where it should not.

This improves e.g., the loading performance and other operations by
a couple percent.