refactor: move unified map into store
This simpliefies the logic of codeMap.preRender.service.

This is a step toward a synchronous and unidirectional data flow model,
decreasing complexitiy and will give us controll of when something
happens or is ready.

It also enables us to extract easily total unary metric in mapTreeViewLevel
component from store, which is needed for its Angular migration.

Refs: #2318