fix: use a debounce while hovering buildings
Hovering a building causes a calculation to draw the name of the
node. Maps with lots of small nodes trigger lots of these
calculations during hovering and that is causing lags. To overcome
the issue, a debounce time is introduced that is pretty much not
noticable by the user while reducing the calculations significantly.

Fixes: https://github.com/MaibornWolff/codecharta/issues/2896