Fix 2896 and 2895 (#2897)
* fix: remove trackEventUsageData to reduce performance overhead

The implementation had a bad time complexity and caused some maps
to take a significant time to load. Remove the tracking as it's
currently not used. We might need that at a later point again. If
that's the case, let's rework the logic.

Fixes: https://github.com/MaibornWolff/codecharta/issues/2895

* fix: use a debounce while hovering buildings

Hovering a building causes a calculation to draw the name of the
node. Maps with lots of small nodes trigger lots of these
calculations during hovering and that is causing lags. To overcome
the issue, a debounce time is introduced that is pretty much not
noticable by the user while reducing the calculations significantly.

Fixes: https://github.com/MaibornWolff/codecharta/issues/2896

* Test usage of debounce in arrow service #2897

* update the changelog
#2897

Co-authored-by: MW-Friedrich <108457762+MW-Friedrich@users.noreply.github.com>
Co-authored-by: Matthias Möser <46388280+knoffi@users.noreply.github.com>
Co-authored-by: knoffi <matthiastutor@outlook.de>