Refactor/2318/migrate download button (#2973)
please not, that downloading and uploading of a map results in a loss of nodes (exact same behaviour as in main). As the download button will be removed in the future, we decided to not spent further time into this. But this is not related to the Angular migration for sure. #2704 might be related.

ref #2318