fix web demo file creation script
The git ls-files call did not create a list of files as it's supposed to, because it look for files relative to the current shell position unlike git log which looks for the project root. Executing the ls-files call in the project root fixes the succeding gitlogparser call and such the problems of the script occuring because it encountered an empty project file.