feat(config): add commitlint to the project to verify commit messages
A commit message always starts with a <type> which can be one of the usual keywords.
The type is followed by an optional <scope>. The scope is encased in braces.
Now follows the subject line, which is in imperative mood written and separated from the colon with a space.
The type and scope need to be lower-case the (here written) body can be any-case